Update README.md
# Coding Landscape Survey

Collect valuable insights from students in the coding community with our survey form. Contribute to shaping the future of coding education by sharing your experiences and preferences. This project is open for contributions and welcomes your input to create a diverse and informative resource.

## How to Contribute

1. Fork the repository.
2. Make changes or improvements.
3. Submit a pull request.

Issues and feedback are appreciated. No specific license is provided; all rights are reserved.

Happy coding!